Update StarLike application command and domain event

+++ only4-KSP-cap4j-application/src/main/java/com/only4/application/commands/star_like/LikeStarCmd.java
@@ -59,10 +59,10 @@ public class LikeStarCmd {
 
 
         @Positive
-        Integer userId;
+        Long userId;
         @Positive
         //TODO:@StarExists
-        Integer starId;
+        Long starId;
     }
 
     /**
