+++ only4-KSP-cap4j-domain/src/main/java/com/only4/domain/aggregates/star_like/factory/StarLikeFactory.java
@@ -38,8 +38,7 @@ public class StarLikeFactory implements AggregateFactory<StarLikeFactory.Payload
     @NoArgsConstructor
     @AllArgsConstructor
     public static class Payload implements AggregatePayload<StarLike> {
-        Integer userId;
-        Integer starId;
-
+        Long userId;
+        Long starId;
     }
 }
